fix(Overlay): Updating to latest version of CDK dependency
This update of npm stays within the given selectors. The package.json
has been updated, but stays within the rules that were already set.
Only packages that are available and are within the current rules were
updated. Run ncu for details on which newer packages are available but
not updated.